adjust event form
