+++ src/styles/globals.css
@@ -21,8 +21,3 @@ main {
 .vertical-line {
     border-left: 2px solid hsl(var(--bc));
 }
-
-.chakra-divider {
-    margin-top: 4em !important;
-    margin-bottom: 4em !important;
-}
